release: promote dev/dc2025 to main (2026.3.0) (#2240)
* disable datatool from scan manager

* uc2000 laser

* fixed ack

* fixed ack

* fixed ack

* fixed ack

* fixed ack

* fixed ack

* fixed ack

* fixed ack

* set laser power hook

* set laser power hook

* fix set laser power

* fix set laser power

* fix set laser power

* fix set laser power

* set laser power 0 on enable

* started eurotherm818

* make metadata repo

* Formatting changes

* started eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated eurotherm818

* updated cryo manager

* Formatting changes

* lazy import get_stage_map_names

* Formatting changes

* lazy import LaserStageMap

* furnace updates

* furnace updates

* furnace updates

* added syn_extraction to experiment editor

* Formatting changes

* fixed XX(CDD) conditionals

* fixed XX(CDD) conditionals

* Formatting changes

* jan hot fixes syn extraction

* Formatting changes

* added saving baseline modifiers

* Formatting changes

* added saving baseline modifiers

* Formatting changes

* jan hot fixes syn extraction

* Formatting changes

* fixed experiment timeseries

* added next_params/next_position

* Formatting changes

* jan hot fixes syn extraction

* Formatting changes

* felix hot fix. make syn_extraction optional

* Formatting changes

* felix hot fix. handle _get_result exception

* Formatting changes

* added timings to dvc persister

* Formatting changes

* jan hot fixes syn extraction, next_position and dynamic pumping

* Formatting changes

* jan hot fixes syn extraction

* Formatting changes

* added modifier_error to load_baselines

* Formatting changes

* updated handling auto linear vs parabolic

* fixes for pyside6. ensure githost loaded before dvc

* Formatting changes

* removed sqlalchemy relation

* pyside6 fix

* fixed display controller

* fixed  handler/controller.init

* Formatting changes

* fixed handler/controller.init

* Formatting changes

* addressed issue #2231

* fixed determine fit

* added updates to zaber_motion_controller

* correction affine

* correction affine

* correction affine

* fixed uc2000

* added power setpoint to controls

* use rebase with sync_repo

* added upload database functionality

* Formatting changes

* added upload database functionality

* Formatting changes

* added upload database functionality

* Formatting changes

* disable print_figure

* Formatting changes

* database upload improvement

* Formatting changes

* database upload improvement

* Formatting changes

* fixed sync repo

* Formatting changes

* fixed sync repo

* use separate data collection branches

* Formatting changes

* added Henry hack

* Formatting changes

* added Henry hack

* Formatting changes

* added power setpoint to controls

* use a single data_collection branch

* cleanup after Henry repo issue

* Formatting changes

* fixed uc2000

* fixed uc2000

* fixed uc2000

* fixed uc2000

* fixed uc2000

* fixed uc2000 move to position

* fixed use_full_power

* disable verbose queries

* fixed reston furnace test_connection

* fixed reston furnace test_connection

* fixed reston furnace test_connection

* fixed reston furnace

* fixed signal calculator

* Formatting changes

* sampleprep tweak

* Formatting changes

* base magnet update

* update about_dialog

* update about_dialog

* Formatting changes

* fixed messurement script

* Formatting changes

* exception handling for ngx lock release

* catch lock release errors

* catch lock release errors

* catch lock release errors

* catch lock release errors

* no lock

* lock

* lock release

* lock release

* lock release

* lock release

* lock release

* lock release

* lock release

* lock release

* lock release

* lock release

* started seed db

* Formatting changes

* added identifer, runid to measurement script
added port to modbustcp_communicator.py
fixed trigger_release

* Formatting changes

* added error to regression editor

* use SAB instead of buffer

* added bakeout plugin

* Formatting changes

* added NGX SAB

* added power law discrimination

* Formatting changes

* hot fixes. removed hardware dependencies from loading task

* Formatting changes

* hot fixes. fixed command register

* added activate/prepare destroy to all managers

* fixed base task activated

* fixed is clean

* Formatting changes

* enable beam blanking for ngx

* Formatting changes

* handle extra kwargs sent to set_mass

* added set to lascon_controller.py

* fixed spectrum

* dc cleanup

* Formatting changes

* updated 2d flux visualization

* Formatting changes

* fixed smart filter

* dc cleanup

* Formatting changes

* sample entry use activate

* qt updates

* updated _enable_laser

* use uv

* refactor: clean up code and remove unnecessary print statements

* feat: specify supported platforms in pyproject.toml

* feat: update pyproject.toml to specify required environments and upgrade pyqt dependency

* fix: update import paths to reflect changes in traitsui and pyface structure

* refactor: replace Color with PyfaceColor in multiple files

* fix: update insert policy handling in combobox_editor.py for compatibility

* fix: add compatibility patch for QDialog exec method in __init__.py

* fix: ensure QDialog exec method compatibility by applying patch in __init__.py

* fix: improve QDialog exec method compatibility by defining a wrapper function in __init__.py

* fix: add patch for Qt IgnoreAction handling in __init__.py

* fix: update Python version requirement to 3.12 and add jinja2 dependency in pyproject.toml

* fix: update Python version requirement to allow any 3.12.x version in pyproject.toml

* fix: normalize detector color input in base_spectrometer.py

* melbourne misc fixes

* fix options color persistence

* misc improvements

* fix rendering irradiation tray canvas. fixed tabular editor colors

* fixed advanced editor area and info indicator

* feat: add AusGeochem EarthData integration with upload functionality

* feat: implement age summary and data point payload builders in earthdata_service

* misc ui improvements

* scan manager fix

* fixed labnumber selection

* chromium co2 laser scan functionality

* ideogram color fix

* fixed colors, fixed stats recalculate

* misc

* fixed flux editor. fixed axis editor

* feat: update pyproject.toml for versioning and dependency constraints

* feat: add MySQL backup and restore scripts for Windows and Linux

* feat: add Z: drive backup script for Windows and update README

* feat: enhance directory handling for repository and meta repository paths

* feat: improve handling of absolute and home directory paths for meta repository

* fix added measured position

* git improvements for dir ownership

* repo: improve ci, workflow, experiment, and pyscript hygiene (#2239)

* ci: make release checks target changed files and focused tests

* ci: fetch base refs before changed-file formatting checks

* ci: skip formatting gate on release promotion branches

---------

+++ pychron/core/ui/qt/video_component_editor.py
@@ -47,7 +47,7 @@ class _VideoComponentEditor(_LaserComponentEditor):
         # self.control.connect(self.playTimer, SIGNAL('timeout()'), self.update)
         print("fffffffffff", self.value.fps)
         if self.value.fps:
-            self.playTimer.setInterval(1000 / float(self.value.fps))
+            self.playTimer.setInterval(int(1000 / float(self.value.fps)))
         self.playTimer.start()
 
         self.value.on_trait_change(self.stop, "closed_event")
@@ -60,7 +60,7 @@ class _VideoComponentEditor(_LaserComponentEditor):
 
     def _update_fps(self):
         if self.value.fps:
-            self.playTimer.setInterval(1000 / float(self.value.fps))
+            self.playTimer.setInterval(int(1000 / float(self.value.fps)))
 
     # def _get_interval(self):
     # if self.value.fps:
